style: 반응형 구현 및 수정사항 변경
