--- src/assets/scss/MyItem.scss
@@ -1,58 +1,0 @@
-.my-item-shop {
-  display: flex;
-  justify-content: center;
-  flex-direction: column;
-  align-items: center;
-
-  .my-item {
-    color: $text-black;
-    font-size: 16px;
-    font-weight: 600;
-    width: 696px;
-    margin-bottom: 16px;
-    left: 0;
-  }
-  .item-content {
-    background-color: $text-white-900;
-    border-radius: 25px;
-    width: 696px;
-
-    .content-box {
-      padding: 20px 28px;
-      display: flex;
-      flex-direction: column;
-
-      .content-title {
-        color: $text-black;
-        font-weight: 600;
-        font-size: 14px;
-        padding-bottom: 10px;
-      }
-      .content-comment {
-        color: $text-gray-900;
-        font-weight: 400;
-        border: none;
-        background-color: $text-white-600;
-        font-size: 17px;
-      }
-    }
-  }
-}
-
-.content-file {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-.add-file {
-  border: 1px solid $brand-blue;
-  color: $brand-blue;
-  border-radius: 8px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 31px;
-  width: 76px;
-  margin: 30px 28px 26px;
-  cursor: pointer;
-}
