+++ src/assets/scss/Myshop.scss
@@ -1,0 +1,38 @@
+.my-item-shop {
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+  align-items: center;
+
+  .my-item {
+    color: $text-black;
+    font-size: 16px;
+    font-weight: 600;
+    margin-bottom: 16px;
+    left: 0;
+  }
+
+  .item-content {
+    background-color: $text-white-900;
+    border-radius: 25px;
+  }
+}
+@media (min-width: 1200px) {
+  .item-content,
+  .my-item {
+    width: 696px;
+  }
+}
+@media (max-width: 1199px) {
+  .item-content,
+  .my-item {
+    width: 696px;
+  }
+}
+
+@media (max-width: 767px) {
+  .item-content,
+  .my-item {
+    width: 344px;
+  }
+}
