+++ src/assets/scss/Styles.scss
@@ -14,4 +14,6 @@
 @import './ShopProfile/ShopProfile.scss';
 @import './ShopLike/ShopLike.scss';
 @import '../scss/ProFileDetail.scss';
-@import './myItem.scss';
+@import '../scss/Myshop.scss';
+@import '../scss/ShopImg.scss';
+@import '../scss/ShopInfo.scss';
